Redux HypedWord refactoring. Epic data fetch fixed
