+++ src/components/Trends/reducers.js
@@ -23,10 +23,9 @@ export default (state = initialState, action) => {
         items: []
       }
     case actions.TRENDS_HYPED_WORD_SELECTED:
-      const { selected } = action.payload
       return {
         ...state,
-        selected
+        selected: action.payload
       }
     default:
       return state
