+++ src/epics/rootEpics.js
@@ -26,7 +26,7 @@ import handleNightModeToggle from './handleNightModeToggle'
 import handleBetaModeToggle from './handleBetaModeToggle'
 import {
   fetchHypedTrends,
-  selectHypedTrends
+  selectHypedTrend
 } from './../components/Trends/fetchHypedTrends'
 import { fetchWordContextEpic } from './../components/WordCloud/fetchWordContextEpic'
 import keyboardEpic from './keyboardEpic'
@@ -60,7 +60,7 @@ export default combineEpics(
   fetchTimeseriesEpic,
   // trends
   fetchHypedTrends,
-  selectHypedTrends,
+  selectHypedTrend,
   // WordCloud
   fetchWordContextEpic
 )
